Fix coverity for wifi wpa supplicant example
[Description]
Casting strlen(buf) from unsigned int to int without checking its value may result in
lost or misinterpreted data.

[Root Cause]
The size_t type has a larger range than the int type; using an int type variable to
store a size_t type variable may result in truncation.

[Fix]
use size_t to store.

[Testing]
compile ok

[Reviewers]
Gannicus Guo, Supriya Sumeet Inamdar, Qingling Wu

[Port proposal]
non-general

+++ freertos/src/supp_main.c
@@ -1037,7 +1037,7 @@ static int hostapd_driver_init(struct hostapd_iface *iface)
     }
 
     /* Initialize the driver interface */
-    if (0U == (b[0] | b[1] | b[2] | b[3] | b[4] | b[5]))
+    if (0U == (unsigned int)(b[0] | b[1] | b[2] | b[3] | b[4] | b[5]))
     {
         b = NULL;
     }
